Improved command handling
The DTMF command objects (LinkCmd) are now handled outside of the Logic class.
They are created in the LinkManager when a logic calls the addLogic function.
The command objects have also been made aware of which link they belong to
which make the command execution simpler since we does not have to figure
that out every time.

Also renamed a few types and variables to more intuitive names.


git-svn-id: svn://svn.code.sf.net/p/svxlink/svn/branches/logiclinking@2397 f8843c6b-282e-0410-804b-c0d5895947d2