Connection calculation code reviewed
The connection calculation code has been rewritten. The old code did work well
but was unnecessarily complicated. Now STL set operations are used to find
out which connections should be established/broken when a link is
activated/deactivated. A number of functions could be removed.

Improved the consistency in naming. A single connection between two logics are
simply called a connection. A group of connections are called a link. A link
is activated or deactivated (not connected/disconnected). Configuration
variable names have not been changed yet though.


git-svn-id: svn://svn.code.sf.net/p/svxlink/svn/branches/logiclinking@2399 f8843c6b-282e-0410-804b-c0d5895947d2