Transmitter shutdown feature added
Now possible to force the transmitter off using a DTMF command. The new
configuration variable SHUTDOWN_CMD is used in a logic configuration section to
set the command that should be used. The node will continue to operate as usual
with the only difference that the transmitter will not be turned on for any
event.


git-svn-id: svn://svn.code.sf.net/p/svxlink/svn/trunk@2404 f8843c6b-282e-0410-804b-c0d5895947d2