Improved the logic shutdown feature
Upon shutdown, any active module is deactivated immediately and module
activation is not allowed until the logic is activated again. This for example
means that if the logic is shut down and an EchoLink connection comes in, that
connection will be rejected. Also, no other DTMF commands than the logic
activation command is allowed when the logic is shut down.

The command was also renamed to ONLINE_CMD since everything became more
logical. Calling it SHUTDOWN_CMD was a bit confusing. Should subcommand 1 mean
that the logic is shut down or brought back online?  Naming the command
ONLINE_CMD make everything clear: 1=online, 0=offline.


git-svn-id: svn://svn.code.sf.net/p/svxlink/svn/trunk@2431 f8843c6b-282e-0410-804b-c0d5895947d2