Goertzel cleanup
The Goertzel algorithm implementation now is correct after verifying it
using the "fft" and "goertzel" commands in MATLAB. Both the C implementation
and the MATLAB code now give exactly the same results.

The previous implementation had a phase problem so the phase was wrong but
the magnitude was correct.

The class documentation has also been improved.


git-svn-id: svn://svn.code.sf.net/p/svxlink/svn/trunk@2436 f8843c6b-282e-0410-804b-c0d5895947d2