Added support for the Opus audio codec
The Opus audio codec is a new audio codec that is superior to all existing free
codecs at bitrates above 6kbps. It also beat a lot of commercial codecs. The
new codec is at the moment only supported with remote receivers, not with
EchoLink.
For more information about Opus, go to http://www.opus-codec.org/.


git-svn-id: svn://svn.code.sf.net/p/svxlink/svn/trunk@2447 f8843c6b-282e-0410-804b-c0d5895947d2