Sound clips for module MetarInfo
The Heather sound clips that Adi contributed a while back have now been tested
and included in the standard sound clip archive. I'll put the archive up on
SourceForge with the next release of SvxLink.

A couple of clips were missing (altimeter and wind) in the originally
contributed archive so they were generated and included in the final archive.
One clip (cavok) was regenerated with improved pronunciation.

The filter_sounds and play_sound scripts were updated to work with newer
versions of 'sox'. This will probably make the scripts fail with older
versions of sox. If that happens, change the SOX_RAW_SAMP_FMT in both
scripts like this:

SOX_RAW_SAMP_FMT="-traw -s2"

This will restore the scripts to the previous behaviour.


git-svn-id: svn://svn.code.sf.net/p/svxlink/svn/trunk@2460 f8843c6b-282e-0410-804b-c0d5895947d2