More modular implementation of PTT hardware support
Support for a specific type of PTT hardware is now contained in its own file.
Both code for serial port pin and GPIO PTT have been removed from the LocalTx
class.

The PTT hardware type must now be specified using the PTT_TYPE configuration
variable in a TX section. Specify "SerialPin" for using a pin in the serial
port, "GPIO" to use a pin in a GPIO port. Set PTT_TYPE to "Dummy" or "NONE" to
not use any PTT hardware at all. It is an error to not specify PTT_TYPE.


git-svn-id: svn://svn.code.sf.net/p/svxlink/svn/trunk@2476 f8843c6b-282e-0410-804b-c0d5895947d2