ModuleEchoLink: New TCL event handling function "squelch_open"
The EchoLink module now have a new TCL event handling function, squelch_open,
which is called when the local receiver squelch open and closes. This function
can be used to for example send a roger beep to the remote station when the
squelch closes.
Patch contributed by DL1HRC / Adi.


git-svn-id: svn://svn.code.sf.net/p/svxlink/svn/trunk@2499 f8843c6b-282e-0410-804b-c0d5895947d2