Block size adjust in Async::AudioDeviceUDP
Had to adjust the block size in Async::AudioDeviceUDP so that the block length
in milliseconds is an integer number. This is necessary since the timer class,
which is used to pace the output, only work with integer milliseconds. If this
adjustment is not made, the audio playback will be a bit too fast.


git-svn-id: svn://svn.code.sf.net/p/svxlink/svn/trunk@2522 f8843c6b-282e-0410-804b-c0d5895947d2