Doxygen documentation can now be built using make
The building of Doxygen documentation have now been set up in the CMake
build system. For this there is a new "doc" target. The building of
compressed manual pages now also are only built if the "doc" target is
invoked explicitly. It is also possible to build one specific type of
documentation. The following targets can be used:

  make doc
  make echolib-doc
  make async-doc
  make man