New configuration variable RAW_AUDIO_UDP_DEST
A new configuration variable for local receivers have been added,
RAW_AUDIO_UDP_DEST, which makes it possible to stream the raw audio read
by SvxLink from the audio device to a UDP port. This feature was mainly
added for debugging but may be usable for other things as well.