Merged the devcal branch into develop
The devcal branch has been used to develop a new calibration utility,
devcal. This utility can be used to calibrate all receivers and
transmitters in the system so that all levels are the same. The default
calibration is done so that max deviation of the channel, e.g. 5kHz, is
-6dBF (half way from maximum AD/DA value). This leaves some headroom for
overdeviation without distorsion in the digital domain.

While implementing this the following also changed:

* New TX config variable: MASTER_GAIN. Use this configuration variable
  to adjust the gain for a transmitter if the slider cannot be pushed
  high enough in the audio mixer or to fine tune during calibration.

* Fractionals can now be used in the PREAMP config var. It was not
  enough with integer resolution to achieve exact calibration.