Fix coverity scan issues 78970, 78983, 78984
Unchecked return value from library -- The function returns a value that
indicates an error condition. If this is not checked, the error
condition may not be handled correctly.
In Async::TcpServer::TcpServer, onConnection: Value returned from a
library function is not checked for errors before being used. This value
may indicate an error condition. (CWE-252)