Ignore coverity scan issue 79041
Time of check time of use -- A check occurs on a file's attributes
before the file is used in a privileged operation, but things may have
changed (CWE-367). An attacker could change the filename's file
association or other attributes between the check and use.

However, in this context, this is a false positive so the issue is
ignored.