AudioSplitter is now a source
The AudioSplitter now also is a source itself so that it can be used to
pipe audio through to the next stage in the pipe instead of having to
use addSink to register all sinks. This make the splitter easier to
handle in an audio pipe where there is a main path and you just want to
split off a branch.