Change default twist threshold for DTMF decoder
The twist thresholds for the DTMF decoder was changed to use 8.5dB for
normal (forward) twist and 6dB for reverse twist. The normal twist
threshold was changed to account for tranceivers using incorrectly
implemented DTMF generators that do not apply pre-emphasis on the
generated tones. Example of such tranceivers are the Yaesu VX-3 and
VX-6. The VX-8 is ok. The reverse twist threshold was changed since it
seemed a bit too liberal. Normal twist should be much more likely than
reverse twist.