Split the RtlTcp class in two
The RtlTcp class was split in two classes, RtlTcp and RtlSdr. The latter
class contain all code that is not specific for how to communicate with
the dongle. The former class (RtlTcp) only contain the specific code
needed to communicate to rtl_tcp via TCP.

This was done to enable the implementation of direct USB communication
with the RTL dongle.