Devcal improvements
The devcal utility can now use wideband FM to do carrier frequency error
calibration. To accomodate this, wideband FM was implemented in the Ddr.

The printouts have also been improved and some helpful instructions was
added.

In rx calibration and measure mode, received audio can now be played
back to an audio device, default alsa:default.

Longer filter times have been applied to the calibration and measurement
printouts. This require the signal to be on for about three seconds to
stabilize.