Fix Coverity scan issue 153979, calling risky function
Coverity scan will warn for uses of "rand" since it's not wise to use
that function in security related contexts. In this case however it was
used to generate noise, so it wasn't a bug really. Changed the call to
rand_r which Coverity scan does not complain about, even if it's the
same thing. It's better to use rand_r anyway since it's thread safe.