Improve flawed reimplementation of DNS lookup
Pull request #202 implement a variant on DNS lookup in
AsyncCppDnsLookupWorker. That reimplementation had a couple of flaws
that this commit should fix.

- The new implementation would not sort out IPv6 answers and produced a
lot of duplicates due to getaddrinfo returning results for multiple
protocols and socket types.
- The new implementation may have had data races after removing the
mutexes. A pthread_join before accessing the shared data should take
care of this problem.
- Improved error/warning messages