Add injectDtmf TCL function
New TCL command, injectDtmf, to inject DTMF digits into the command
stream just as if they were received on the receiver. This function can
for example be used in conjuction with the dtmf_digit_received and
dtmf_cmd_received functions to remap commands to something different or
inventing compound commands.