Split base of Logic class into new class LogicBase
The bare necessity of the Logic class have now been split out into the
new class LogicBase. That class only have the most basic functionality
for a logic. The only thing it can do is work with the link manager to
connect to other logic cores and stream audio into and out of the logic.
The LogicBase class can be used to implement all sorts of audio sources
and sinks and not just transceiver based logic cores.