Adjust voting delay for slow receivers
When configuring a voter using the RECEIVERS configuration variable,
optionally the receiver name may be followed by a colon and a squelch
delay value. This can be used to adjust the voting delay for receivers
that always are slower to open for some reason. The delay value given
will be subtracted from the voting delay if the receiver is the first
one to open.