Add error checking code to Async::AudioProcessor
The Async::AudioProcessor could get into an invalid state if a connected
audio sink returned an invalid sample count. This commit add a check for
the return value.
Also added a couple of debug printouts to find out if any bad return
values are present.

This change was made to adress issue #209.