Use a frequency sampling filter for OB AFSK
Use a frequency sampling filter as the filter for out of band AFSK. This
filter suppress the speech band much better than the previous filtering
but at the same time it's more computationally efficient.

Also remove the ability to set AFSK parameters like, center frequency,
shift and baudrate in the configuration file.