Fix memory hadling bug in ModuleMetarInfo
Memory was freed while still in use. Function onActivity called
ClearWatchMap which erased all FdWatches. That also erased the signal
objects that called the onActivity function. Upon return from onActivity
to the destroyed signal object the module crashed the whole application.
This did not happed on all platforms but typically on Raspberry Pi.