The reflector server can now handle newer clients
The reflector server now have a mechanism to handle clients using a newer
protocol version than the server supports. If a client upon connect
indicates a protocol version newer than what the server can handle it
will ask the client to downgrade its protocol version. It is up to the
client to choose wether it can handle the downgrade or not.