Bugfix: RX not signalling sql open after MUTE_ALL
After disabling a receiver (MUTE_ALL) and then enabling it again, if the
squelch was open at both operations, a squelch open would not be
signalled after enabling it. This at least caused problems if using
RemoteTrx and then in SvxLink was using the DISABLE/ENABLE PTY
functionality in the receiver voter.

Closes #610