New configuration variable TG_SELECT_INHIBIT_TIMEOUT
A new configuration variable TG_SELECT_INHIBIT_TIMEOUT has been added
that is used to separately control the TG select inhibit feature.

Talkgroup selection is inhibited when there is activity on the local frequency
that does not activate a talkgroup. Only activation due to remote activity
(monitoring) is inhibited. A talkgroup can be activated as usual by using any
method for local activation. This feature make it possible to conduct a local
QSO without being disturbed by remote reflector traffic.

Use this configuration variable to set the number of seconds of inactivity
before the node will go back to monitoring remote talkgroup activity. The
default value is taken from TG_SELECT_TIMEOUT if this variable is not set.
Setting a value of 0 will disable the feature entirely.