Write to SvxLink Server logfile from a thread
Writing to the logfile in SvxLink Server is now done from a separate
thread.  That will increase the accuracy of the timestamps in some
situations.  Previously multiple log rows could be lumped together in a
block where all rows had the same timestamp. That made it harder to
debug some problems.