APRS: If LocationInfo TONE is not set, Toff is sent
Previously, when TONE was not set in the LocationInfo configuration, the
tone was sent as T000 in the APRS packet. Now Toff is sent instead which
is compliant with the standard.

Also a check was added so that TONE can only be set between 0 and 9999.
It is the responsibility of the sysop to set a value that is valid
within that range.