APRS: Use timestamp 111111z for EL-objects
Go back to send Echolink object with the special timestamp 111111z.
This unique time stamp declares this object to be permanent.  This means
that it should not be replaced by any other similarly named object
unless that object is transmitted by the same originating station. This
lets the originator of a permanent object update or move his object, but
it then prevents his object from being replaced by anyone else's similar
object.